tools/requirements.txt
pillow==10.3.0
GitPython==3.1.32
